+++ src/example/ExamplePage.vue
@@ -18,6 +18,9 @@
       <li class="nav-item">
         <router-link to="/decimal" class="nav-link">Decimal</router-link>
       </li>
+      <li class="nav-item">
+        <router-link to="/unique" class="nav-link">Unique</router-link>
+      </li>
     </ul>
 
     <div>